requirements.txt
Flask==2.2.3
Pillow==9.4.0
transformers==4.26.1
